flask-app-passkey/requirements.txt
Flask
Authlib
requests
PyJWT 